Revert "Navigation refactor & MVP Story/Chapter Views"
