+++ app/src/App.vue
@@ -1,9 +1,16 @@
 <template>
   <v-app id="app">
     <v-content>
-      <transition name="component-fade" mode="out-in">
-        <router-view/>
-      </transition>
+      <v-container fluid >
+        <v-layout align-center justify-center>
+          <v-flex xs12 sm8 md4>
+            <v-img :src="require('@/assets/logo.png')" height="128px" contain class="d-block my-4"></v-img>
+            <transition name="component-fade" mode="out-in">
+              <router-view/>
+            </transition>
+          </v-flex>
+        </v-layout>
+      </v-container>
     </v-content>
     <v-snackbar v-model="notification" :multi-line="true" :timeout="3000">
       {{ notificationText }}
