--- app/src/chapters.js
@@ -1,133 +1,0 @@
-export default [
-        {
-          title: "Meet the Piggles",
-          folder: "introduction",
-          thumbnail: "thumbnail.png",
-          mediaType: "video",
-          videoName: "introduction.mp4",
-          length: "02:51",
-          released: true
-        },
-        {
-          title: "Chapter 1",
-          folder: "chapter1",
-          thumbnail: "thumbnail.png",
-          blurb: "Get to know the Piggles. I this adventure....",
-          readTime: [5, 7],
-          mediaType: "book",
-          released: false,
-          pages: [
-            {
-              image: "Intro image 1.png",
-              text: "‘Whose idea was it to move in the middle of winter?’, Penny Piggle wondered, as her breath frosted the glass of the piggy-mini van window."
-            },
-            {
-              image: "Intro image 2.png",
-              text: "They'd been packing for weeks. All Penny and her brother Rex knew was that they were going to a bigger place."
-            },
-            {
-              image: "Intro image 3.png",
-              text: "Penny and Rex’s parents couldn’t wait to show them in person. They wanted to keep everything about their new house a surprise, so they didn’t show the piglets any pictures."
-            },
-            {
-              image: "Intro image 4.png",
-              text: "Penny sat with her snout against the glass, squinting to see through the falling ice and snow. ‘How did their new house look?’, she worried. 'Would she like it?', she thought."
-            },
-            {
-              image: "Intro image 5.png",
-              text: "It felt like they'd been driving forever. Finally, Mama Piggle turned around. \"Look! We're almost to our new house! Would you two like a clue?\"."
-            },
-          ]
-        },
-        {
-          title: "Chapter2",
-          folder: "chapter2",
-          thumbnail: "thumbnail.png",
-          blurb: "Get to know the Piggles. I this adventure....",
-          readTime: [5, 7], 
-          mediaType: "book",
-          released: false,
-          pages: [
-            {
-              image: "Intro image 1.png",
-              text: "Those great big fluffy clouds. We artists are a different breed of people. We're a happy bunch. We don't have to be committed. We are just playing here."
-            },
-            {
-              image: "Intro image 2.png",
-              text: "Here's another little happy bush That's why I paint - because I can create the kind of world I want - and I can make this world as happy as I want it. You can spend all day playing with mountains. We'll put a happy little sky in here. You need the dark in order to show the light. You have to make those little noises or it won't work."
-            },
-            {
-              image: "Intro image 3.png",
-              text: "Penny and Rex’s parents couldn’t wait to show them in person. They wanted to keep everything about their new house a surprise, so they didn’t show the piglets any pictures."
-            },
-            {
-              image: "Intro image 4.png",
-              text: "Penny sat with her snout against the glass, squinting to see through the falling ice and snow. ‘How did their new house look?’, she worried. 'Would she like it?', she thought."
-            },
-            {
-              image: "Intro image 5.png",
-              text: "It felt like they'd been driving forever. Finally, Mama Piggle turned around. \"Look! We're almost to our new house! Would you two like a clue?\"."
-            },
-          ]
-        },
-        // {
-        //   title: "Introduction 3",
-        //   folder: "chapter3",
-        //   thumbnail: "thumbnail.png",
-        //   blurb: "Get to know the Piggles. I this adventure....",
-        //   readTime: [5, 7], 
-        //   mediaType: "book",
-        //   pages: [
-        //     {
-        //       image: "Intro image 1.png",
-        //       text: "‘Whose idea was it to move in the middle of winter?’, Penny Piggle wondered, as her breath frosted the glass of the piggy-mini van window."
-        //     },
-        //     {
-        //       image: "Intro image 2.png",
-        //       text: "They'd been packing for weeks. All Penny and her brother Rex knew was that they were going to a bigger place."
-        //     },
-        //     {
-        //       image: "Intro image 3.png",
-        //       text: "I like to beat the brush. Now we can begin working on lots of happy little things. Play with the angles. Clouds are free they come and go as they please."
-        //     },
-        //     {
-        //       image: "Intro image 4.png",
-        //       text: "Use absolutely no pressure. Just like an angel's wing. You could sit here for weeks with your one hair brush trying to do that - or you could do it with one stroke with an almighty brush. Even the worst thing we can do here is good."
-        //     },
-        //     {
-        //       image: "Intro image 5.png",
-        //       text: "You're the greatest thing that has ever been or ever will be. You're special. You're so very special. You don't have to spend all your time thinking about what you're doing, you just let it happen. Mountains are so simple, they're hard. You can create anything that makes you happy. Isn't that fantastic that you can make whole mountains in minutes?"
-        //     },
-        //   ]
-        // },
-        // {
-        //   title: "Introduction 4",
-        //   folder: "chapter4",
-        //   thumbnail: "thumbnail.png",
-        //   blurb: "Get to know the Piggles. I this adventure....",
-        //   readTime: [5, 7], 
-        //   mediaType: "book",
-        //   pages: [
-        //     {
-        //       image: "Intro image 1.png",
-        //       text: "‘Whose idea was it to move in the middle of winter?’, Penny Piggle wondered, as her breath frosted the glass of the piggy-mini van window."
-        //     },
-        //     {
-        //       image: "Intro image 2.png",
-        //       text: "They'd been packing for weeks. All Penny and her brother Rex knew was that they were going to a bigger place."
-        //     },
-        //     {
-        //       image: "Intro image 3.png",
-        //       text: "Penny and Rex’s parents couldn’t wait to show them in person. They wanted to keep everything about their new house a surprise, so they didn’t show the piglets any pictures."
-        //     },
-        //     {
-        //       image: "Intro image 4.png",
-        //       text: "Penny sat with her snout against the glass, squinting to see through the falling ice and snow. ‘How did their new house look?’, she worried. 'Would she like it?', she thought."
-        //     },
-        //     {
-        //       image: "Intro image 5.png",
-        //       text: "It felt like they'd been driving forever. Finally, Mama Piggle turned around. \"Look! We're almost to our new house! Would you two like a clue?\"."
-        //     },
-        //   ]
-        // },
-      ]
